refactor: ingredient details
- remove duplication of useParams
- use clean ingredientDetails in inredient page
- remove withModal since modal forms from url, not ingredient click
