--- src/components/burger-ingredients/burger-ingredient/burger-ingredient-with-modal.jsx
@@ -1,8 +1,0 @@
-import { withModal } from "../../../hocs/withModal";
-import { BugerIngredient } from "./burger-ingredient";
-import { IngredientDetailsModal } from "../../ingredient-details/ingredient-details-modal";
-
-export const BurgerIngredientWithModal = withModal(
-  BugerIngredient,
-  IngredientDetailsModal,
-);
