+++ src/components/ingredient-details/index.js
@@ -1,2 +1,3 @@
 export { IngredientDetailsParams } from "./ingredient-details-params";
 export { IngredientDetailsModal } from "./ingredient-details-modal";
+export { IngredientDetails } from "./ingredient-details";
